backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pydantic-settings==2.7.1
python-dotenv==1.0.1
python-multipart==0.0.20
--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.0
torchvision>=0.15
transformers>=4.30
timm>=0.9
Pillow>=10.0
opencv-python-headless>=4.8
numpy>=1.24
python-jose[cryptography]>=3.3
bcrypt>=4.0.1
pytest>=8.0
httpx>=0.27
